requirements.txt
#fastapi dependencies
fastapi
uvicorn
python-dotenv
pydantic
pydantic_core
# langchain dependencies, # for testing purposes
langchain
langchain-community
langchain-openai
# langchain-huggingface
# langchain-chroma
langchain-core
langchain-prompts
langgraph
pinecone
langchain_pinecone
langchain_upstage #this shitter needs cl.exe (c/c++ compiler for windows) to install
# chromadb
# faiss-cpu
# sentence-transformers
openai
# huggingface-hub
# unstructured
# unstructured[local-inference]
# unstructured[all-docs]
